// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const CATEGORY = {
  ALL: 'ALL',
  OFFERS: 'OFFERS',
  ORGANIC_HONEY: 'ORGANIC_HONEY',
  HONEY_MIXES: 'HONEY_MIXES',
  HIVE_PRODUCTS: 'HIVE_PRODUCTS',
  FARM_PRODUCTS: 'FARM_PRODUCTS'
} as const

export type CATEGORY = (typeof CATEGORY)[keyof typeof CATEGORY]